# Optimize warmup for large sites