--- a/changes/CA-3625.change
+++ b/changes/CA-3625.change
@@ -1,0 +1,1 @@
+Optimize warmup: reduce warmup time and memory consumption after warmup. [buchi]