Loading the most important catalog indexes into memory doesn't work well for larger sites. It takes several minutes and fills up a lot of memory. We try another approach by calling REST API endpoints that are always called by the frontend on a new page load. At least for read access and when using the new frontend this approach seems to be superior.

Test with ogg1m, new approach:

Warmup duration: `24s`
Memory consumption after warmup: `500 MB`
First Page Load after warmup (Dashboard of new frontend): `7.5s`

Test with ogg1m, old approach:

Warmup duration: `5m 13s`
Memory consumption after warmup: `3.4 GB`
First Page Load after warmup (Dashboard of new frontend): `9.7s`

Test Köniz (very slow disk system), new approach:

Warmup duration: `56s`
Memory consumption after warmup: `400 MB`
First Page Load after warmup (Personal overview of classic frontend): `22.45s`

Test Köniz, old approach:

Warmup duration: `7m 11s`
Memory consumption after warmup: `2.2 GB`
First Page Load after warmup (Personal overview of classic frontend): `30.53s`

There are some views in the classic frontend that are significantly slower on the first call with the new approach. E.g. repository overview (`tabbed_view/listing?view_name=overview`). These views can be improved by using Solr.

For [CA-3625](https://4teamwork.atlassian.net/browse/CA-3625)

## Checklist

_Everything has to be done/checked. Checked but not present means the author deemed it unnecessary._

- [x] Changelog entry
- [x] Link to issue (Jira or GitHub) and backlink in issue (Jira)